RetroSignalModel

RetroSignalModel.ONE_SIGNAL

RetroSignalModel.RtgMTK

RetroSignalModel.ZERO_SIGNAL

RetroSignalModel.hil

RetroSignalModel.load_conditions

RetroSignalModel.load_data

RetroSignalModel.load_parameters

RetroSignalModel.resting_u0

RetroSignalModel.scan_params

RetroSignalModel.ONE_SIGNAL — Method

Damage signal = 1

RetroSignalModel.RtgMTK — Function

rtgMTK(S = DEFAULT_SIGNAL; name, proteinlevels=STRESSED)

Make a RTG signalling ModelingToolkit (mtk) system with a time-dependent damage signal function S and proteinlevels for Rtg1, Rtg2, Rtg3, Bmh, and Mks proteins.

RetroSignalModel.ZERO_SIGNAL — Method

Damage signal = 0

RetroSignalModel.hil — Method

hil(x, k, n=1)

Hill function

\[hil(x, k, n) := rac{x^n}{x^n + k^n}\]

RetroSignalModel.load_conditions — Function

Load boolean conditions for model validation.

RetroSignalModel.load_data — Method

Loads a CSV file and returns a RowTable.

RetroSignalModel.load_parameters — Function

Load model parameters

RetroSignalModel.resting_u0 — Method

Resting initial conditions: all zeroes for all active components.

RetroSignalModel.scan_params — Function

Scan for parameters that meet the boolean conditions in the retrograde (RTG) signalling model.