RetroSignalModel

RetroSignalModel.K_dist

RetroSignalModel.SSMETHOD

RetroSignalModel.actM1_model

RetroSignalModel.capVec

RetroSignalModel.construct

RetroSignalModel.getOutput

RetroSignalModel.getSteady

RetroSignalModel.getSteadySol

RetroSignalModel.get_protein_lookup

RetroSignalModel.init_u

RetroSignalModel.isValid

RetroSignalModel.knockout

RetroSignalModel.paramSearching

RetroSignalModel.paramSearching_Once

RetroSignalModel.rtgM1_model

RetroSignalModel.rtgM2_model

RetroSignalModel.rtgM3_model

RetroSignalModel.rtgM4_model

RetroSignalModel.setTuple

RetroSignalModel.setTuples

RetroSignalModel.try_conditions

RetroSignalModel.actM1_model — Method

Activation Model:

RetroSignalModel.capVec — Method

Scale maximum element to capVal. v is a vector

RetroSignalModel.construct — Method

Get constructor from a data type

RetroSignalModel.getOutput — Method

Compare the cytosolic concentration of either 'rtg1' or 'rtg3' with their nucleus concentrations. Output: 1 : nucleus concentration is higher than the cytosolic one 0 : otherwise

RetroSignalModel.getSteady — Method

Returm RTGmodel with u in steady state.

RetroSignalModel.getSteadySol — Method

Solve steady state of RTGmodel based on given initial variables (u). Noted that fieldname u is ignored.

RetroSignalModel.get_protein_lookup — Method

Protein lookup table: reported the indexes of specify protein and location

RetroSignalModel.init_u — Method

Create initial values of reaction systems based on protein total concentration. Each concentration is solved by Linear programming.

RetroSignalModel.isValid — Method

Validate responses

RetroSignalModel.knockout — Method

Knockout specified protein with name prName, and set as low concentration equal to DEL_CONC

RetroSignalModel.paramSearching — Method

Parameter searching Api. It can be run in single core or multiple-processing

Performance

julia> addprocs(exeflags="--project=/home/runner/work/RetroSignalModel.jl/RetroSignalModel.jl/docs/Project.toml");
julia> @time @everywhere using RetroSignalModel
julia> nprocs()
9
julia> @time rs.paramSearching(rs.rtgM4(); num_sim=100, distributed=true, saveall=true)
  2.378919 seconds (78.95 k allocations: 4.964 MiB)

RetroSignalModel.paramSearching_Once — Method

Sampling one parameter set for validation

RetroSignalModel.rtgM1_model — Method

Rtg Model: version 1

RetroSignalModel.rtgM2_model — Method

This model is modified version of rtgM1. Major changes:

Change input to be michaelis menten

Change the translocation to be Michaelis menten

Rtg1p and Rtg3p enter nucleus separately

RetroSignalModel.rtgM3_model — Method

This model is modified version of rtgM2. Major changes:

Change input to be michaelis menten

Change the translocation to be Michaelis menten: (which is not ture). See supplemental material

Rtg1p and Rtg3p enter nucleus separately

References

Hao, N., Budnik, B. A., Gunawardena, J., & O'Shea, E. K. (2013). Tunable signal processing through modular control of transcription factor translocation. Science, 339(6118), 460-464. [Article, SOM]

Mitochondrial Signaling: The Retrograde Response. 2004. (Reaction of Bmh-Mks on Rtg13 dimer) [Article]

RetroSignalModel.rtgM4_model — Method

Modified version from rtgM3

Change input to Hill kinetics

RetroSignalModel.setTuple — Method

Change tuple element and return a new one

RetroSignalModel.setTuples — Method

Change tuple elements and return a new one

RetroSignalModel.try_conditions — Method

Try conditions of the model with given initial variables and parameters.

RetroSignalModel.K_dist — Constant

Parameter searching

RetroSignalModel.SSMETHOD — Constant

Steady state solver.

A Stiff solver is chosen TRBDF2 (stand for Tranpezoidal Backward Differeital Formula) from DifferentialEquations.jl for robust simulation of stability.

TRBF2 is a one-step method based on trapezoidal rule and the backward differentiaion formula of order 2, and it is strongly L-stable that means TRBFS is good at integrating stiff equations [1].

References

Hosea, M. E., & Shampine, L. F. (1996). Analysis and implementation of TR-BDF2. Applied Numerical Mathematics, 20(1-2), 21-37. URL: https://www.sciencedirect.com/science/article/pii/0168927495001158?via%3Dihub

DifferentialEqiations.jl - Stiff Problems. [link]