RetroSignalModel

RetroSignalModel.actM1

RetroSignalModel.rtgM1

RetroSignalModel.rtgM2

RetroSignalModel.rtgM3

RetroSignalModel.Param

RetroSignalModel.Protein

RetroSignalModel.model_meta

RetroSignalModel.model_meta

RetroSignalModel.valid_setting

RetroSignalModel.valid_setting

RetroSignalModel.cal_dudt

RetroSignalModel.dudt

RetroSignalModel.get_Exptable

RetroSignalModel.get_del_u

RetroSignalModel.get_index_dictionary

RetroSignalModel.get_output

RetroSignalModel.get_outputQ

RetroSignalModel.get_protein_lookup

RetroSignalModel.try_conditions

RetroSignalModel.try_simulate

RetroSignalModel.actM1 — Constant

Activation Model:

RetroSignalModel.rtgM1 — Constant

Rtg Model: version 1

RetroSignalModel.rtgM2 — Constant

This model is modified version of rtgM1. Major changes:

Change input to be michaelis menten

Change the translocation to be Michaelis menten

Rtg1p and Rtg3p enter nucleus separately

RetroSignalModel.rtgM3 — Constant

This model is modified version of rtgM2. Major changes:

Change input to be michaelis menten

Change the translocation to be Michaelis menten: (which is not ture). See the supplemental material of https://www.pnas.org/content/pnas/suppl/2006/06/30/0604085103.DC1/04085SuppText.pdf

Rtg1p and Rtg3p enter nucleus separately

Ref:

https://science.sciencemag.org/content/339/6118/460

https://www.pnas.org/content/pnas/suppl/2006/06/30/0604085103.DC1/04085SuppText.pdf

https://www.cell.com/molecular-cell/pdf/S1097-2765(04)00179-0.pdf (Reaction of Bmh-Mks on Rtg13 dimer)

RetroSignalModel.Param — Type

Parameter setting for model

RetroSignalModel.Protein — Type

Compartment concentration of protein

RetroSignalModel.model_meta — Type

Model meta. To store information of model object, conditions, and protein lookup table

RetroSignalModel.model_meta — Method

model_meta(model)

Get model meta. Including the origianl model input, conditional data from csv, and protein lookup table

RetroSignalModel.valid_setting — Type

Setting for validation.

RetroSignalModel.valid_setting — Method

Generate validation setting from setting.jl.

RetroSignalModel.cal_dudt — Function

Return the derivatives of the DE model

RetroSignalModel.dudt — Method

dudt(model, u, p; t=nothing)

Get the derivative in given u and p from Catalyst model.

RetroSignalModel.get_Exptable — Method

Get Boolean model

RetroSignalModel.get_del_u — Method

Copy vector u, and delete elements with lookup table

RetroSignalModel.get_index_dictionary — Method

Get Dictionary of index. Both agents (u) and parameters(p) Example: julia> indexdict = getindexdictionary(MODEL); WARNING: redefining constant indexdict julia> indexdict.u.s 1 julia> indexdict.p.n_s 1

RetroSignalModel.get_output — Method

Compare the cytosolic concentration of either 'rtg1' or 'rtg3' with their nucleus concentrations. Output: 1 : nucleus concentration is higher than the cytosolic one 0 : otherwise

RetroSignalModel.get_outputQ — Method

Qualitative Validation

method

Substraction of nuclear concentration with cytoplasmic concentration

RetroSignalModel.get_protein_lookup — Method

Protein lookup table: reported the indexes of specify protein and location

RetroSignalModel.try_conditions — Method

Try conditions of the model with given initial variables and parameters.

RetroSignalModel.try_simulate — Method

Test for solving ODE.